nix: switch from crate2nix to crane
crate2nix builds one derivation per dependency which is cool and great
if you want to pull in a bunch of separate flakes and have them share
derivations. But it's really complicated and adds a TON of flake
dependencies.

Crane, on the other hand, is very simple:

1. It builds the dependencies.
2. It builds the root crate.

And that's it.

It's not as smart (it doesn't pick up native dependencies), but that's
fine. I can add those manually.